refactor: refactor ui component folder
