+++ src/pages/addCard/components/AddCardForm.tsx
@@ -1,4 +1,4 @@
-import Button from "../../../components/button/Button";
+import Button from "../../../components/ui-kit/Button";
 import useAddCardForm from "../hooks/useAddCardForm";
 import CardExpiredDateInput from "./CardExpiredDateInput";
 import CardHolderNameInput from "./CardHolderNameInput";
