+++ src/pages/addCard/components/CardExpiredDateInput.tsx
@@ -1,4 +1,4 @@
-import { Input, InputProps } from "../../../components/input/Input";
+import { Input, InputProps } from "../../../components/ui-kit/Input";
 
 interface CardExpiredDateInputProps {
   value: string;
