+++ src/pages/addCard/components/CardPasswordInput.tsx
@@ -1,4 +1,4 @@
-import { Input, InputProps } from "../../../components/input/Input";
+import { Input, InputProps } from "../../../components/ui-kit/Input";
 import { GoDotFill } from "react-icons/go";
 import Dot from "./Dot";
