+++ src/pages/addCard/components/Header.tsx
@@ -1,5 +1,5 @@
 import { FaChevronLeft } from "react-icons/fa";
-import Button from "../../../components/button/Button";
+import Button from "../../../components/ui-kit/Button";
 import { useNavigate } from "react-router-dom";
 
 export default function Header() {
