+++ src/pages/cards/components/AddCardButton.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
-import Card from "../../../components/card/Card";
+import Card from "../../../components/ui-kit/Card";
 import { FaPlus } from "react-icons/fa6";
-import Button from "../../../components/button/Button";
+import Button from "../../../components/ui-kit/Button";
 
 export default function AddCardButton() {
   const navigate = useNavigate();
